Fix test naming
