+++ javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/Issue1485Test.java
@@ -18,7 +18,7 @@ import java.nio.file.Path;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class Issue1485 extends AbstractSymbolResolutionTest {
+class Issue1485Test extends AbstractSymbolResolutionTest {
 
     @Test
     void issue1485withoutSpecifyingJARs() throws IOException {
