+++ javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/Issue18Test.java
@@ -19,7 +19,6 @@ package com.github.javaparser.symbolsolver;
 import com.github.javaparser.ast.CompilationUnit;
 import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
 import com.github.javaparser.ast.body.MethodDeclaration;
-import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.stmt.ExpressionStmt;
 import com.github.javaparser.resolution.types.ResolvedType;
 import com.github.javaparser.symbolsolver.javaparser.Navigator;
@@ -31,21 +30,17 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class Issue116 extends AbstractResolutionTest {
+class Issue18Test extends AbstractResolutionTest {
 
     @Test
-    void arrayTypeIsNotPartOfTheTree() {
-        CompilationUnit cu = parseSample("Issue116");
-        ClassOrInterfaceDeclaration clazz = Navigator.demandClass(cu, "JavaTest");
-        MethodDeclaration methodDeclaration = Navigator.demandMethod(clazz, "foo");
+    void typeDeclarationSuperClassImplicitlyIncludeObject() {
+        CompilationUnit cu = parseSample("Issue18");
+        ClassOrInterfaceDeclaration clazz = Navigator.demandClass(cu, "Foo");
+        MethodDeclaration methodDeclaration = Navigator.demandMethod(clazz, "bar");
+        ExpressionStmt expr = (ExpressionStmt) methodDeclaration.getBody().get().getStatements().get(1);
         TypeSolver typeSolver = new ReflectionTypeSolver();
         JavaParserFacade javaParserFacade = JavaParserFacade.get(typeSolver);
-        com.github.javaparser.ast.type.Type typeNode = methodDeclaration.getParameters().get(0).getType();
-        ResolvedType type = javaParserFacade.convert(typeNode, typeNode);
-        assertEquals("java.lang.String[]", type.describe());
-
-        ExpressionStmt expressionStmt = (ExpressionStmt) methodDeclaration.getBody().get().getStatements().get(0);
-        Expression argRef = expressionStmt.getExpression();
-        assertEquals("java.lang.String[]", javaParserFacade.getType(argRef).describe());
+        ResolvedType type = javaParserFacade.getType(expr.getExpression());
+        assertEquals("java.lang.Object", type.describe());
     }
 }
