+++ javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/Issue235Test.java
@@ -17,7 +17,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.Arrays;
 import java.util.Collection;
 
-class Issue235 extends AbstractResolutionTest{
+class Issue235Test extends AbstractResolutionTest{
 
     static Collection<String> data() {
         return Arrays.asList(
