--- javaparser-symbol-solver-testing/src/test/java/com/github/javaparser/symbolsolver/Issue84.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2016 Federico Tomassetti
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.javaparser.symbolsolver;
-
-import com.github.javaparser.ast.CompilationUnit;
-import com.github.javaparser.ast.expr.MethodCallExpr;
-import com.github.javaparser.resolution.types.ResolvedType;
-import com.github.javaparser.symbolsolver.javaparser.Navigator;
-import com.github.javaparser.symbolsolver.javaparsermodel.JavaParserFacade;
-import com.github.javaparser.symbolsolver.resolution.AbstractResolutionTest;
-import com.github.javaparser.symbolsolver.resolution.typesolvers.ReflectionTypeSolver;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class Issue84 extends AbstractResolutionTest {
-
-    @Test
-    void variadicIssue() {
-        CompilationUnit cu = parseSample("Issue84");
-        final MethodCallExpr methodCall = Navigator.findMethodCall(cu, "variadicMethod").get();
-
-        final JavaParserFacade javaParserFacade = JavaParserFacade.get(new ReflectionTypeSolver());
-        final ResolvedType type = javaParserFacade.getType(methodCall);
-        assertEquals(String.class.getCanonicalName(), type.asReferenceType().getQualifiedName());
-    }
-}
